requirement.txt
absl-py==2.2.1
ace_tools==0.0
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
beautifulsoup4==4.13.4
bidict==0.23.1
blinker==1.9.0
blis==0.7.11
bs4==0.0.2
cachetools==6.1.0
catalogue==2.0.10
certifi==2025.6.15
charset-normalizer==2.1.1
click==8.1.8
cloudpathlib==0.21.0
cn-clip==1.5.1
confection==0.1.5
contourpy==1.3.0
cycler==0.12.1
cymem==2.0.11
dgl==2.1.0
distro==1.9.0
exceptiongroup==1.3.0
fake-useragent==2.2.0
filelock==3.13.1
Flask==3.1.1
Flask-SocketIO==5.5.1
flatbuffers==25.2.10
fonttools==4.58.0
fsspec==2024.6.1
ftfy==6.3.1
gast==0.6.0
gensim==4.3.3
google-pasta==0.2.0
greenlet==3.2.3
grpcio==1.71.0
h11==0.16.0
h5py==3.13.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.4
importlib_metadata==8.6.1
importlib_resources==6.5.2
itsdangerous==2.2.0
jieba==0.42.1
Jinja2==3.1.4
jiter==0.10.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
keras==3.9.1
keras-bert==0.89.0
keras-embed-sim==0.10.0
keras-layer-normalization==0.16.0
keras-multi-head==0.29.0
keras-pos-embd==0.13.0
keras-position-wise-feed-forward==0.8.0
keras-self-attention==0.51.0
keras-transformer==0.40.0
kiwisolver==1.4.7
langchain-core==0.3.65
langchain-openai==0.3.24
langcodes==3.5.0
langsmith==0.3.45
language_data==1.3.0
libclang==18.1.1
lmdb==1.3.0
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.4
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
murmurhash==1.0.12
namex==0.0.8
networkx==3.2.1
node2vec==0.5.0
numpy==1.26.4
openai==1.88.0
opt_einsum==3.4.0
optree==0.14.1
orjson==3.10.18
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pip==25.0
playwright==1.52.0
preshed==3.0.9
protobuf==5.29.4
psutil==7.0.0
PuLP==3.1.1
pydantic==2.11.4
pydantic_core==2.33.2
pydantic-settings==2.9.1
pyee==13.0.0
Pygments==2.19.1
PyJWT==2.8.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.1
python-socketio==5.13.0
pytz==2025.2
PyYAML==6.0
ratelimit==2.2.1
regex==2024.11.6
requests==2.28.1
requests-toolbelt==1.0.0
rich==14.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==4.1.0
setuptools==75.8.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.7
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-pkuseg==0.0.33
srsly==2.5.1
sympy==1.13.1
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tf_keras==2.19.0
thinc==8.2.5
threadpoolctl==3.6.0
tiktoken==0.9.0
timm==1.0.15
tokenizers==0.21.1
torch==2.1.0+cu121
torchaudio==2.1.0+cu121
torchdata==0.7.1
torchvision==0.16.0+cu121
tqdm==4.67.1
transformers==4.51.3
triton==2.1.0
typer==0.15.3
typing_extensions==4.12.2
typing-inspection==0.4.0
tzdata==2025.2
urllib3==1.26.13
uv==0.7.6
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
wsproto==1.2.0
xlrd==2.0.2
xxhash==3.5.0
zh-core-web-lg==3.7.0
zh-core-web-sm==3.7.0
zhipuai==2.1.5.20250626
zipp==3.21.0
zstandard==0.23.0